rm unused pages
